+++ app/src/main/java/com/seungma/infratalk/data/datasource/remote/FirebaseLikeRemoteDataSourceImpl.kt
@@ -20,21 +20,6 @@ import java.util.Date
 import javax.inject.Inject
 
 
-interface LikeDataSource {
-    suspend fun insertBoardLike(boardLikeInsertRequest: BoardLikeInsertRequest): LikeResponse
-    suspend fun deleteBoardLike(boardLikeDeleteRequest: BoardLikeDeleteRequest): LikeResponse
-    suspend fun selectBoardLike(boardLikeSelectRequest: BoardLikeSelectRequest): LikeResponse
-    suspend fun selectBoardLikeCount(boardLikeCountSelectRequest: BoardLikeCountSelectRequest): LikeCountResponse
-
-    suspend fun insertCommentLike(commentLikeInsertRequest: CommentLikeInsertRequest): LikeResponse
-    suspend fun deleteCommentLike(commentLikeDeleteRequest: CommentLikeDeleteRequest): LikeResponse
-    suspend fun selectCommentLike(commentLikeSelectRequest: CommentLikeSelectRequest): LikeResponse
-    suspend fun selectCommentLikeCount(commentLikeCountSelectRequest: CommentLikeCountSelectRequest): LikeCountResponse
-    suspend fun deleteCommentRelatedLikes(commentRelatedLikesDeleteRequest: CommentRelatedLikesDeleteRequest): CommentRelatedLikesResponse
-    suspend fun deleteBoardLikes(boardLikesDeleteRequest: BoardLikesDeleteRequest): BoardLikesDeleteResponse
-}
-
-
 class FirebaseLikeRemoteDataSourceImpl @Inject constructor(
     private val database: FirebaseFirestore
 ) : LikeDataSource {
