+++ app/src/main/java/com/seungma/infratalk/data/datasource/remote/ImageDataSource.kt
@@ -1,0 +1,8 @@
+package com.seungma.infratalk.data.datasource.remote
+
+import com.seungma.infratalk.data.model.request.image.ImagesRequest
+import com.seungma.infratalk.data.model.response.image.ImagesResponse
+
+interface ImageDataSource {
+    suspend fun uploadImages(imagesRequest: ImagesRequest): ImagesResponse
+}
